requirements.txt
camelot_py==0.11.0
fitz==0.0.1.dev2
Flask==3.0.2
numpy>=1.21
openai==0.28.0
opencv-python==4.9.0.80
openpyxl==3.0.10
pandas==1.4.4
pdf2image==1.16.3
Pillow==9.2.0
pymupdf==1.23.6 
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
python_Levenshtein==0.25.0
tabula_py==2.9.0
Werkzeug==3.0.1
gunicorn
flask_cors
